Organize Neo4j and Oracle DB schema builders into a 'schemas' directory
